+++ php/school_app/src/Controller/Students.php
@@ -1,6 +1,6 @@
 <?php namespace SchoolApp\Controller;
 
-use \SchoolApp\Model\Person as Student;
+use \SchoolApp\Model\Student as Student;
 use \MongoDB\BSON\ObjectId as ObjectId; 
 
 class Students {
@@ -10,13 +10,12 @@ class Students {
     }
     static function create(Student $student) {
         \SchoolApp\Repository\Students::insertOne($student);
-        return 200;
+        return "successful insert!";
     }
     static function show(string $student) {
         $objectId = new ObjectId($student);
         $found_student = \SchoolApp\Repository\Students::findOne($objectId);
         \SchoolApp\View\Students::index([$found_student]); 
-        return 200;
     }
 }
