+++ php/school_app/src/Model/Student.php
@@ -1,0 +1,22 @@
+<?php namespace SchoolApp\Model;
+
+class Student {
+
+    public function __construct(string $name, array $courses ){
+        $this->name = $name;
+        $this->courses = $courses;
+    }
+
+    public static function makeWithPost(array $post) {
+        return new self($post['name'], $post['courses']);
+    }
+
+    function get(): array {
+        return [
+            'name' => $this->name,
+            'courses' => $this->courses,
+        ];
+    }
+}
+
+?>
